cxx-frontend: remove dependency on TreeTransform + ASTMutator.cpp content
