--- tools/mull-cxx-frontend/src/MullTreeTransform.cpp
@@ -1,5 +1,0 @@
-//
-// Created by Stanislav Pankevich on 23.04.21.
-//
-
-#include "MullTreeTransform.h"
